Alpha 2.0.2
After maintenance following changes added:
(~) Should have fixed the bug of page and/or image not showing up problems. Was due to GitHub not updating.
(-) Removed folders which are currently within the file structure that not in use, they are now safely stored.

All files have been re-added and not modified unless explained otherwise in the update log.