Version 5.6
(+) Added the Geneva Conventions Purity Test
(+) Added a list of all communes in the Vaud canton, alongside with maps and flags. This page is unlisted as the links are pasted from wikipedia so it's intended for strictly personal use. There is also a pdf version of the page on the website, also unlisted.
(~) Put the Pantheons up to date (haven't done much to the Gigachad Pantheon though)
(-) Removed the games folder on the website. Everything else will just fit in the "other category." Thus, the OGVR section has been moved to Other and will probably be available in the next update.
(-) Removed the standard pages (e.g 403, 451, blank page, etc.). No need for them.