Eleventh Version
(+) Added the French version for a good amount of pages. Links on the French version to pages which aren't (yet) supported in French just lead to the English version. An -fr suffix is added to each translated page's URL, OR the page is just renamed to the French version. This is also an experiment to see how different translations work together. The last page I want translated it the one pertaining to local history, but that shall come in the next version.
(~) Redid some parts of the index; added the this is where I post from image.
(~) Changed the naming convention of the updates. It makes me feel too much of a STEM kid to call this version 6.0. Big changes with additions are referred to as Versions, small tweaks are Updates within the Version. An equivalent name has been given to each Version and Update below. If pages are added to the website (or something just as big) then it's a new Version. If it's a little modification like a bug fix, fixing a spelling mistake, or updating a Pantheon and nothing more, then it's just an Update.
(-) Removed superfluous crap and made the website whose pages were translated more concise in general.
I'll probably be taking a break from updating the website now. It's in a place I'd like it to be. Expect modifications relating to OGVR and local history when I come back.