V11U1
(+) Added a law section to the website.
(+) Added a the index of all useful legal resources.
(+) Added the Canevas de Recherche Juridique Informatisée, a useful resource for the upcoming exam.
(~) Note: the code for versions will change now - this one is coded as V11U1 on GitHub. V = version. U = Update.
(~) Second note: extremely minor changes like PDF uploads will not be mentioned. Modifications to the links on the new link page will not be added to the change log unless its notification is deemed necessary. In the GitHub repository such changes will be marked with the code TDDMMYYYYHHMM - T for tweak, d for day, m for month, etc.